Land sign out on a dedicated page instead of redirecting straight off the dashboard
Sign out used to run entirely from the click handler: revoke the token,
build the RP-Initiated Logout URL, then hand off to a location.href change,
all while the dashboard sat rendered underneath. It also pointed
post_logout_redirect_uri at the console root, so the OP's redirect back
landed on a protected route and ProtectedRoute immediately kicked off a
fresh sign-in, chaining another cross-document navigation.

Sign out now navigates to a new unprotected /signing-out route first (a
plain client-side transition, no flash) that shows a spinner and performs
the actual sign-out request, and post_logout_redirect_uri now points back
at that same route so the OP's redirect lands there instead of the
dashboard. The page checks the session on mount and moves on to home once
it's confirmed clear.
