+++ frontend/apps/console/src/layouts/__tests__/DashboardLayout.test.tsx
@@ -1,16 +1,11 @@
 // Copyright 2025 The ThunderID Authors
 // SPDX-License-Identifier: Apache-2.0
 
-import {render, screen, userEvent, waitFor} from '@thunderid/test-utils';
-import {afterEach, describe, it, expect, vi, beforeEach} from 'vitest';
+import {render, screen, userEvent} from '@thunderid/test-utils';
+import {describe, it, expect, vi, beforeEach} from 'vitest';
 import DashboardLayout from '../DashboardLayout';
 
 const mockNavigate = vi.fn();
-const mockSignIn = vi.fn();
-const mockSignOut = vi.fn();
-const mockClearSession = vi.fn();
-const mockLoggerError = vi.fn();
-const mockLoggerWarn = vi.fn();
 const mockUserData = vi.fn();
 interface MockUseGetApplicationsResult {
   data?: {
@@ -24,8 +19,6 @@ interface MockUseGetApplicationsResult {
 }
 
 const mockUseGetApplications = vi.fn<(params: unknown) => MockUseGetApplicationsResult>();
-let mockDiscovery: {wellKnown?: {end_session_endpoint?: string}} | undefined;
-let mockIsTrustedIssuerGenericOidc = false;
 
 vi.mock('@thunderid/configure-applications', async (importOriginal) => ({
   ...(await importOriginal<typeof import('@thunderid/configure-applications')>()),
@@ -34,14 +27,7 @@ vi.mock('@thunderid/configure-applications', async (importOriginal) => ({
 
 // Mock ThunderID
 vi.mock('@thunderid/react', () => ({
-  useThunderID: () => ({
-    signIn: mockSignIn,
-    clearSession: mockClearSession,
-    discovery: mockDiscovery,
-  }),
   User: ({children}: {children: (user: unknown) => React.ReactNode}) => children(mockUserData()),
-  SignOutButton: ({children}: {children: (props: {signOut: () => void}) => React.ReactNode}) =>
-    children({signOut: mockSignOut}),
 }));
 
 // Mock contexts
@@ -57,9 +43,6 @@ vi.mock('@thunderid/contexts', async (importOriginal) => {
         },
         client: {client_id: 'CONSOLE'},
       },
-      isTrustedIssuerGenericOidc: () => mockIsTrustedIssuerGenericOidc,
-      getTrustedIssuerClientId: () => 'test-client-id',
-      getClientUrl: () => 'https://localhost:5191/console',
     }),
   };
 });
@@ -68,16 +51,6 @@ vi.mock('@thunderid/contexts', async (importOriginal) => {
 vi.mock('react-i18next', () => ({
   useTranslation: () => ({
     t: (key: string) => key,
-  }),
-}));
-
-// Mock logger
-vi.mock('@thunderid/logger/react', () => ({
-  useLogger: () => ({
-    error: mockLoggerError,
-    warn: mockLoggerWarn,
-    info: vi.fn(),
-    debug: vi.fn(),
   }),
 }));
 
@@ -101,8 +74,6 @@ describe('DashboardLayout', () => {
     vi.clearAllMocks();
     sessionStorage.clear();
     mockUserData.mockReturnValue({name: 'Test User', email: 'test@example.com'});
-    mockIsTrustedIssuerGenericOidc = false;
-    mockDiscovery = undefined;
     mockUseGetApplications.mockReturnValue({
       data: {applications: []},
       isLoading: false,
@@ -150,9 +121,8 @@ describe('DashboardLayout', () => {
     expect(screen.getByText(new RegExp(currentYear.toString()))).toBeInTheDocument();
   });
 
-  it('does not start a second sign-in after signing out', async () => {
+  it('navigates to the signing-out page when sign out is clicked', async () => {
     const user = userEvent.setup();
-    mockSignOut.mockResolvedValue(undefined);
 
     render(<DashboardLayout />);
 
@@ -164,31 +134,7 @@ describe('DashboardLayout', () => {
     const signOutButton = await screen.findByText('common:userMenu.signOut');
     await user.click(signOutButton);
 
-    await waitFor(() => {
-      expect(mockSignOut).toHaveBeenCalled();
-    });
-    expect(mockSignIn).not.toHaveBeenCalled();
-  });
-
-  it('logs error when signOut fails', async () => {
-    const user = userEvent.setup();
-    const signOutError = new Error('Sign out failed');
-    mockSignOut.mockRejectedValue(signOutError);
-
-    render(<DashboardLayout />);
-
-    // Open the user menu first
-    const userMenuTrigger = screen.getByLabelText('Test User');
-    await user.click(userMenuTrigger);
-
-    // Click sign out menu item
-    const signOutButton = await screen.findByText('common:userMenu.signOut');
-    await user.click(signOutButton);
-
-    await waitFor(() => {
-      expect(mockSignOut).toHaveBeenCalled();
-      expect(mockLoggerError).toHaveBeenCalledWith('Sign out failed', {error: signOutError});
-    });
+    expect(mockNavigate).toHaveBeenCalledWith('/signing-out');
   });
 
   it('renders the user profile picture in the account menu when available', () => {
@@ -232,82 +178,5 @@ describe('DashboardLayout', () => {
     const welcomeItem = await screen.findByText('common:userMenu.welcome');
     expect(welcomeItem).toBeInTheDocument();
     await user.click(welcomeItem);
-  });
-
-  describe('generic OIDC sign out', () => {
-    let originalLocation: Location;
-
-    beforeEach(() => {
-      mockIsTrustedIssuerGenericOidc = true;
-      originalLocation = window.location;
-      Object.defineProperty(window, 'location', {
-        value: {...originalLocation, href: ''},
-        writable: true,
-        configurable: true,
-      });
-    });
-
-    afterEach(() => {
-      Object.defineProperty(window, 'location', {
-        value: originalLocation,
-        writable: true,
-        configurable: true,
-      });
-    });
-
-    it('clears local session and redirects to client URL when end_session_endpoint is missing', async () => {
-      mockDiscovery = {wellKnown: {}};
-      const user = userEvent.setup();
-
-      render(<DashboardLayout />);
-
-      const userMenuTrigger = screen.getByLabelText('Test User');
-      await user.click(userMenuTrigger);
-
-      const signOutButton = await screen.findByText('common:userMenu.signOut');
-      await user.click(signOutButton);
-
-      expect(mockClearSession).toHaveBeenCalled();
-      expect(mockLoggerWarn).toHaveBeenCalledWith(expect.stringContaining('end_session_endpoint missing'));
-      expect(window.location.href).toBe('https://localhost:5191/console');
-    });
-
-    it('clears local session and redirects to IdP end_session_endpoint when available', async () => {
-      mockDiscovery = {wellKnown: {end_session_endpoint: 'https://idp.example.com/logout'}};
-      const user = userEvent.setup();
-
-      render(<DashboardLayout />);
-
-      const userMenuTrigger = screen.getByLabelText('Test User');
-      await user.click(userMenuTrigger);
-
-      const signOutButton = await screen.findByText('common:userMenu.signOut');
-      await user.click(signOutButton);
-
-      expect(mockClearSession).toHaveBeenCalled();
-      expect(window.location.href).toContain('https://idp.example.com/logout');
-      expect(window.location.href).toContain('client_id=test-client-id');
-    });
-
-    it('logs error when clearSession throws during generic OIDC sign out', async () => {
-      mockDiscovery = {wellKnown: {end_session_endpoint: 'https://idp.example.com/logout'}};
-      const sessionError = new Error('session clear failed');
-      mockClearSession.mockImplementation(() => {
-        throw sessionError;
-      });
-      const user = userEvent.setup();
-
-      render(<DashboardLayout />);
-
-      const userMenuTrigger = screen.getByLabelText('Test User');
-      await user.click(userMenuTrigger);
-
-      const signOutButton = await screen.findByText('common:userMenu.signOut');
-      await user.click(signOutButton);
-
-      expect(mockLoggerError).toHaveBeenCalledWith(expect.stringContaining('Failed to clear local session'), {
-        error: sessionError,
-      });
-    });
   });
 });
